pyproject.toml

[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["dbp"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
freezegun = "^1.4.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
